@@ -9,10 +9,10 @@
   "main": "src/index.js",
   "repository": {
     "type": "git",
-    "url": "https://github.com/"
+    "url": "https://github.com/awen2841/babel-plugin-react-if"
   },
   "bugs": {
-    "url": "https://github.com//issues"
+    "url": "https://github.com/awen2841/babel-plugin-react-if/issues"
   },
   "keywords": [
     "react",
@@ -29,27 +29,34 @@
   ],
   "license": "MIT",
   "scripts": {
-    "lint": "eslint src spec",
+    "lint": "eslint src spec *.js --ignore-pattern '!*.js'",
     "test": "mocha spec/*.js",
-    "cover": "istanbul cover -x \"**/spec/**\" node_modules/mocha/bin/_mocha spec/*.js"
+    "cover": "istanbul cover -x \"**/spec/**\" node_modules/mocha/bin/_mocha spec/*.js",
+    "upgrade": "ncu -u && npm i",
+    "postversion": "npm publish && git push && git push --tags"
   },
   "dependencies": {
-    "babel-core": "^6.1.2",
-    "babel-plugin-syntax-jsx": "^6.1.18"
+    "@babel/plugin-syntax-jsx": "^7.16.7",
+    "@babel/template": "^7.16.7"
   },
   "devDependencies": {
-    "babel-eslint": "^6.1.2",
-    "babel-preset-react": "^6.1.2",
-    "chai": "^3.5.0",
-    "chai-enzyme": "^0.5.1",
-    "coveralls": "^2.11.6",
-    "enzyme": "^2.4.1",
-    "eslint": "^3.4.0",
-    "eslint-config-airbnb": "^10.0.1",
-    "eslint-plugin-react": "^6.2.0",
-    "istanbul": "^0.4.1",
-    "mocha": "^3.0.2",
-    "react": "^15.3.x",
-    "react-dom": "^15.3.x"
+    "@babel/core": "^7.17.7",
+    "@babel/preset-react": "^7.16.7",
+    "@babel/register": "^7.17.7",
+    "babel-eslint": "^10.1.0",
+    "chai": "^4.3.6",
+    "chai-enzyme": "^1.0.0-beta.1",
+    "coveralls": "^3.1.1",
+    "enzyme": "^3.11.0",
+    "enzyme-adapter-react-16": "^1.15.6",
+    "eslint": "^7.32.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-plugin-import": "^2.25.4",
+    "eslint-plugin-react": "^7.29.4",
+    "istanbul": "^0.4.5",
+    "mocha": "^9.2.2",
+    "react": "^17.0.x",
+    "react-dom": "^17.0.x",
+    "npm-check-updates": "^12.5.3"
   }
 }
