@@ -1,9 +1,9 @@
-var React = require('react');
+const React = require('react');
 
-module.exports = React.createClass({
-  render: function() {
+module.exports = class extends React.Component {
+  render() {
     return (
       <button>{this.props.children}</button>
     );
   }
-});
+};
