@@ -1,9 +1,13 @@
-var RIF = require('../src/index');
+const Enzyme = require('enzyme');
+const Adapter = require('enzyme-adapter-react-16');
+const RIF = require('../src/index');
 
-require('babel-core/register')({
-  presets: ['babel-preset-react'],
+Enzyme.configure({ adapter: new Adapter() });
+
+require('@babel/register')({
+  presets: ['@babel/preset-react'],
   plugins: [RIF],
-  cache: false
+  cache: false,
 });
 
 require('./test/r-if.test');
