@@ -1,36 +1,38 @@
-var getChildren = require('./util/getChildren');
-var childrenHaveAttributes = require('./util/childrenHaveAttributes');
-var getJSXElement = require('./util/getJSXElement');
-var getAttributes = require('./util/getAttributes');
-var getAttribute = require('./util/getAttribute');
-var transformChildren = require('./util/transformChildren');
-var getAttributeConditionExpression = require('./util/getAttributeConditionExpression');
-var attributeNames = require('./util/constants');
+const getChildren = require('./util/getChildren');
+const childrenHaveAttributes = require('./util/childrenHaveAttributes');
+const getJSXElement = require('./util/getJSXElement');
+const getAttributes = require('./util/getAttributes');
+const getAttribute = require('./util/getAttribute');
+const transformChildren = require('./util/transformChildren');
+const getAttributeConditionExpression = require('./util/getAttributeConditionExpression');
+const attributeNames = require('./util/constants');
 
 module.exports = function (babel) {
-  var visitor = {
-    JSXElement: function (path) {
-      var attributes = getAttributes(path.node);
+  const visitor = {
+    JSXElement(path) {
+      const attributes = getAttributes(path.node);
 
-      var children = getChildren(babel.types, path.node);
+      const children = getChildren(babel.types, path.node);
 
       if (!childrenHaveAttributes(
         babel.types,
         children,
-        [attributeNames.R_IF, attributeNames.R_ELSE_IF, attributeNames.R_ELSE]
+        [attributeNames.R_IF, attributeNames.R_ELSE_IF, attributeNames.R_ELSE],
       )) {
-        var attributeWithIf = getAttribute(
+        const attributeWithIf = getAttribute(
           babel.types,
           attributes,
-          attributeNames.R_IF
+          attributeNames.R_IF,
         );
 
         if (!attributeWithIf) {
           return;
         }
 
-        var existsConditionExpression = getAttributeConditionExpression(
-          babel.types, attributeNames.R_IF, attributeWithIf
+        const existsConditionExpression = getAttributeConditionExpression(
+          babel.types,
+          attributeNames.R_IF,
+          attributeWithIf,
         );
 
         path.replaceWithMultiple(
@@ -41,14 +43,14 @@ module.exports = function (babel) {
               path.node,
               attributes.filter((attr) => attr.name.name !== attributeNames.R_IF),
             ),
-            babel.types.NullLiteral()
-          )
+            babel.types.NullLiteral(),
+          ),
         );
 
         return;
       }
 
-      var newChildren = transformChildren(babel.types, children);
+      const newChildren = transformChildren(babel.types, children);
 
       path.replaceWithMultiple([
         getJSXElement(
@@ -56,13 +58,13 @@ module.exports = function (babel) {
           path.node,
           attributes,
           newChildren,
-        )
+        ),
       ]);
-    }
+    },
   };
 
   return {
-    inherits: require('babel-plugin-syntax-jsx'),
-    visitor: visitor
+    inherits: require('@babel/plugin-syntax-jsx').default,
+    visitor,
   };
 };
