@@ -1,9 +1,9 @@
-var getAttributes = require('./getAttributes');
-var includesAttributeName = require('./includesAttributeName');
+const getAttributes = require('./getAttributes');
+const includesAttributeName = require('./includesAttributeName');
 
 function childrenHaveAttributes(types, children, names) {
-  return children.some(node => {
-    var attributes = getAttributes(node);
+  return children.some((node) => {
+    const attributes = getAttributes(node);
 
     return includesAttributeName(types, attributes, names);
   });
