@@ -1,7 +1,5 @@
-
-
 module.exports = {
   R_IF: 'r-if',
   R_ELSE_IF: 'r-else-if',
-  R_ELSE: 'r-else'
+  R_ELSE: 'r-else',
 };
